Create a huge mess trying to watch for changes
Had some fun actually trying to write Rust here but I realised at the last
minute most of this hackery is totally unnecessary.

What we need at most is to parse out all of the refs that are referenced in a
range spec. The details of how they all relate to commits is not important,
e.g. we don't care if the user specified "foo" or "^foo@{yesterday}". We just
need to observe when the ref "foo" is modified, then pass the full range back
to git rev-parse again.

In fact, maybe we don't need any of that crap - perhaps all we need is to watch
the whole .git dir for changes of any relevant type, regardless of the files
they touch. Then we can just call git rev-list again.