WIP: Attempt to test resource module with loom
I _think_ that the tokio Semaphore automatically works with Loom but not sure.

However this doesn't work because async_condvar_fair doesn't work with Loom.

I tried to follow
https://docs.rs/async-condvar-fair/latest/async_condvar_fair/macro.RelockMutexGuard.html#third-party-impl-inconvenient-mutex
to make it work but it doesnn't seem to work because the macro seems to require
that the guard type is Send and std::sync::Mutex is not.

At this point I think this is too complicated to adopt, not worth it.