WIP - tired. Trying to create general notification mechanism
Because the status tracker will need to know which tests are _gonna_ be run, we need a richer communication with the test manager. I'm trying to implement this by extending the results notifications to just be general test status notifications.

Currently just need to work through the build errors and see where we end up.