Add LCI_CONFIG_DIR (BROKEN)
I was originally thinking to use this for packing up scripts and config
files alongside the main local-ci.toml. However I realised that's not
really much good unless we also hash those config files and stuff. That
would be a much more advanced kinda feature...

Anyway this is currently broken because it just takes the config path
as-is without checking if it's relative.