MORE WIP
I think maybe all we have to do is just tidy this up a bit and then it's fine? Not sure.

I didn't have to change set_revisions very much to make it....... not insanely
big, it's just a bit ugly and verbose because of the way Rust just is, I think?
At least with my skills not sure I can make it any more beautiful.

Next step is to try and clean up the notification logic a bit, and then see if
we can get await_job_success looking a bit less loco.