Add failing test cases for dag::tests
The logic in the implementation of BottomUp and TopDown is faulty:
in graphs with diamond configurations, the iterators duplicate nodes
and don't respect the topological order.

This commit adds two unit tests which would fail, but they are
commented out so that CI doesn't break. Next commit will fix the bugs
and enable the tests.